refactor(RandomPointsGenerator): 도메인 패키지로 이동
